% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_size.R
\name{block_size}
\alias{block_size}
\title{Calculate the size of spatial blocks to use for data partitioning based on the estimation of spatial autocorrelation}
\usage{
block_size(points, raster, num_sample, crs)
}
\arguments{
\item{points}{A list of point datasets, each in a data.frame format with two columns in the order of "long" and "lat"}

\item{raster}{RasterStack of environmental data to use for calculation}

\item{num_sample}{Integer. The number of sample points to use. See ?cv_spatial_autocor documentation of the blockCV package for details}

\item{crs}{A coordinate reference system (CRS). Use the 4 digit numbers of EPSG codes}
}
\value{
A list of numbers indicating the size of spatial blocks to use
}
\description{
Calculate the size of spatial blocks to use for data partitioning based on the estimation of spatial autocorrelation
}
\examples{
get.block.size <- block_size(points = list(bg1, bg2, bg3), raster = envs, num_sample = 5000, crs = 4326)
}
