% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_data.R
\name{box_data}
\alias{box_data}
\title{Miscellaneous tool to aid data visualization: format environmental data extracted from rasters to make boxplots in ggplot2}
\usage{
box_data(sp.name, envs, pts)
}
\arguments{
\item{sp.name}{Charcter. Input species name}

\item{envs}{RasterStack of environmental layers}

\item{pts}{A data.frame of species occurrence points having two columns, in the order of "long" and "lat"}
}
\value{
A data.frame of formatted for use in ggplot2
}
\description{
Miscellaneous tool to aid data visualization: format environmental data extracted from rasters to make boxplots in ggplot2
}
\examples{
get.box.data <- box_data(sp.name = 'Karsenia koreana', envs = envs, pts = occs)
}
