% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffMaker.R
\name{buffMaker}
\alias{buffMaker}
\title{Generate spatial buffers for input occurrence points}
\usage{
buffMaker(occs_list, envs, buff_dist)
}
\arguments{
\item{occs_list}{A list containing occurrence datasets. Each occurrence set needs to have two columns named "long" and "lat", and in that order}

\item{envs}{RasterStack  (or RasterLayer) of input environmental data}

\item{buff_dist}{A numeric value given in meters. Circular dissolved buffers will be created based on this value}
}
\value{
A list of spatial buffers, with the order of species corresponding to the input species occurrence points
}
\description{
Generate spatial buffers for input occurrence points
}
\examples{
get.buff <- buffMaker(occs_list = occs_list, envs = envs, buff_dist = 700000)
}
