% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pred_diff.R
\name{calc_pred_diff}
\alias{calc_pred_diff}
\title{Calculate the difference between ENM predictions}
\usage{
calc_pred_diff(layer.a, layer.b, method, verbose = F)
}
\arguments{
\item{layer.a}{RasterLayer of first prediction map}

\item{layer.b}{RasterLayer of second prediction map}

\item{verbose}{print out messages when method = 'Schoener' is used.}

\item{type}{The specific method of calculating prediction differences. If method = 'Dubos', conduct calculation based on the equation provided in Dubos et al. (2023).
If method = 'Schoener', use the ENMTools package to first calculate the Schoener's D metric, and then use the equation (1 - D) * 100 to calculate the difference between predictions}
}
\value{
A percent value of difference between the two input RasterLayers
}
\description{
Calculate the difference between ENM predictions
}
\examples{
diff <- calc_pred_diff(layer.a = pred.a, layer.b = pred.b, method = 'Dubos')
}
