% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ranges.R
\name{calc_ranges}
\alias{calc_ranges}
\title{Calculate the range size of suitable habitats from binary presence/absence maps}
\usage{
calc_ranges(bin.stack, bin.labs, digits)
}
\arguments{
\item{bin.stack}{RasterStack object containing presence/absence maps converted from continuous ENM predictions}

\item{bin.labs}{Vector of layer names corresponding to the presence/absence layers}

\item{digits}{Integer value used to round the results of range calculation}
}
\value{
A list where each slot contains the calculated range value for each presence/absence map
}
\description{
Calculate the range size of suitable habitats from binary presence/absence maps
}
\examples{
get_ranges <- calc_ranges(bin.stack = bin, bin.labs = names(bin), digits = 0)
}
