% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_mess.R
<<<<<<< HEAD:man/do_mess.Rd
\name{do_mess}
\alias{do_mess}
=======
\name{do.mess}
\alias{do.mess}
>>>>>>> 2bf5a1856620003319edb856ffcebb5fc19a03b0:man/do.mess.Rd
\title{Conduct Multivariate Environmental Similiarity Surfaces (MESS) analysis across different projection scenarios}
\usage{
do_mess(proj.env, proj.names, ref.env, occs)
}
\arguments{
\item{proj.env}{RasterStack object of the porjction data}

\item{proj.names}{A vector of layer names for the projection data}

\item{ref.env}{RasterStack object of reference (= calibration) data}

\item{occs}{A data.frame of species occurrence points containing two columns, with the "long" column followed by "lat" column}
}
\value{
RasterStack of MESS layer for each projection scenario
}
\description{
Conduct Multivariate Environmental Similiarity Surfaces (MESS) analysis across different projection scenarios
}
\examples{
get.mess <- do.mess(proj.env = proj.stack, proj.names = c('LGM', 'Current', '2050'), ref.env = ref.stack, occs = occs)
}
