% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bias_grid.R
\name{get_bias_grid}
\alias{get_bias_grid}
\title{Generate a bias grid to use for spatial sampling bias correction}
\usage{
get_bias_grid(targ.pts, envs, poly)
}
\arguments{
\item{targ.pts}{A data.frame of target group occurrence points containing two columns, in the order of "long" and "lat"}

\item{envs}{RasterStack of environmental data}

\item{poly}{SpatialPolygonsDataFrame object to mask the output bias grid}
}
\value{
RasterLayer of generated bias grid
}
\description{
Generate a bias grid to use for spatial sampling bias correction
}
\examples{
bias <- get_bias_grid(targ.pts = targ.pts, envs = envs, poly = poly)
}
