% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_occs_list.R
\name{make_occs_list}
\alias{make_occs_list}
\title{Format occurrence dataframe into a list for occurrence thinning}
\usage{
make_occs_list(occs_df, spp_list, long_col, lat_col)
}
\arguments{
\item{occs_df}{Dataframe of occurrences merged across species}

\item{spp_list}{List of species names}

\item{long_col}{Character. Name of longitude column}

\item{lat_col}{Character. Name of latitude column}
}
\description{
Format occurrence dataframe into a list for occurrence thinning
}
\examples{
occs_list <- make_occs_list(occs_df = occs_df, spp_list = spp_list, long_col = 'long', latitude = 'lat')
}
