% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occs_thinner.R
\name{occs_thinner}
\alias{occs_thinner}
\title{Spatial thinning of (multiple) species occurrence datasets, using the SDMtune package (Vignali et al. 2020)}
\usage{
occs_thinner(occs_list, envs, long, lat, spp_list)
}
\arguments{
\item{occs_list}{A list of data.frame objects for species occurrence points. Need to have "long" and "lat" columns}

\item{envs}{RasterLayer. The spatial resolution of this layer will be used as a spatial thinning distance}

\item{long}{Character. Name of the longitude column}

\item{lat}{Character. Name of the latitude column}

\item{spp_list}{A list of input species names}
}
\value{
A list of dataframes containing the thinned occurrence points for each input species
}
\description{
Spatial thinning of (multiple) species occurrence datasets, using the SDMtune package (Vignali et al. 2020)
}
\examples{
thin_occs <- occs_thinner(occs_list = occs_list, envs = envs, long = 'long', lat = 'lat', spp_list = spp_list)
}
