% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_preds.R
\name{plot_preds}
\alias{plot_preds}
\title{Plot model prediction outputs}
\usage{
plot_preds(preds, poly, ncol = NULL, nrow = NULL, colors, high, low)
}
\arguments{
\item{preds}{RasterStack object of model prediction outputs}

\item{poly}{A polygon shapefile of area boundaries to overlay on top of the prediction rasters}

\item{ncol}{(optional) the number of columns to use in the plotting pane}

\item{nrow}{(optional) the number of rows to use in the plotting pane}

\item{colors}{A color palette to use for plotting}

\item{high}{Numeric. A higher-end value of estimated suitability}

\item{low}{Numeric. A lower-end value of estimated suitability}
}
\value{
A ggplot2-style plot of model prediction outputs
}
\description{
Plot model prediction outputs
}
\examples{
plot_current <- plot_preds(preds = preds, poly = poly, ncol = 2, nrow = 3, colors = c('#2b83ba', '#abdda4', '#ffffbf', '#fdae61', '#4f05d7'))
}
