% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resp_data_pull.R
<<<<<<< HEAD:man/resp_data_pull.Rd
\name{resp_data_pull}
\alias{resp_data_pull}
=======
\name{respDataPull}
\alias{respDataPull}
>>>>>>> 2bf5a1856620003319edb856ffcebb5fc19a03b0:man/respDataPull.Rd
\title{Extract MaxEnt response curve data for further plot customization}
\usage{
resp_data_pull(sp.name, model, names.var)
}
\arguments{
\item{sp.name}{Character. The name of the taxon being used in niche modeling}

\item{model}{A fitted ENM of the class MaxEnt. Can be extracted by accessing the "models" slot output from "test_multibg" or "test_multisp"}

\item{names.var}{Character. Names of environmental variables used to generate the MaxEnt model}
}
\value{
A data.frame containing the response data that can be used to further customize the response curve plots
}
\description{
Extract MaxEnt response curve data for further plot customization
}
\examples{
resp.data <- respDataPull(sp.name = 'Karsenia koreana', model = mx_model, var = c('bio1', 'bio12', 'slope'))
}
