% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_multibg.R
\name{test_multibg}
\alias{test_multibg}
\title{MaxEnt model tuning with multiple background datasets}
\usage{
test_multibg(
  taxon.name,
  occs,
  envs,
  bg.list,
  tune.args,
  partitions,
  partition.settings = NULL,
  user.grp = NULL,
  type
)
}
\arguments{
\item{taxon.name}{Character. The name of the taxon being used in niche modeling}

\item{occs}{Data frame of species occurrence points with two columns, in the order of "long" and "lat"}

\item{envs}{RasterStack of environmental data}

\item{bg.list}{A list of background points to test within the modeling tuning workflow. Each background set should be in a two-column data.frame format with column names matching the occurrence dataset}

\item{tune.args}{A named list containing feature class settings ("fc") and regularization values ("rm"). Use ?ENMevaluate for details}

\item{partitions}{Character. The name of data partitioning strategy to use for model evaluations. Can select from: "randomkfold", "block", "checkerboard1", "checkerboard2", "testing", and "user"}

\item{partition.settings}{A named list used to specify settings for the selected partitioning scheme}

\item{user.grp}{A named list designating user-specified cross-validation folds. Only required when "partition.settings" is set to "user"}

\item{type}{Character. Select a method for selecting the optimal parameter combinations. If "type1" is used, the optimal model is selected based on the minimum 10% omission rate as a primary criterion, minimum AUCDIFF value as a secondary criterion, and maximum validation AUC as a tertiary criterion
If "type2" is used, the primary, secondary, and tertiary criteria for model selection would be delta.AICc smaller than 2, minimum 10% omission rate, and maximum validation AUC, respectively}
}
\value{
A named list with slots for model metrics of selected models, model objects, model predictions, and variable importance
}
\description{
MaxEnt model tuning with multiple background datasets
}
\examples{
test_bg <- test_multibg(taxon.name = 'Karsenia koreana', occs = occs, envs = envs, bg.list(bg1, bg2, bg3), tune.args = tune.args, partitions = 'randomkfold', partition.settings = list(kfolds = 10), type = 'type1')
}
