% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_multisp.R
\name{test_multisp}
\alias{test_multisp}
\title{MaxEnt model tuning for multiple species}
\usage{
test_multisp(
  taxon.list,
  occs.list,
  envs,
  bg,
  tune.args,
  partitions,
  partition.settings = NULL,
  user.grp = NULL,
  type
)
}
\arguments{
\item{taxon.list}{A vector of species names to run model tuning}

\item{occs.list}{A list containing occurrence points for each species. The occurrence dataset for each species should be a data.frame object with two columns, in the order of longitude and latitude}

\item{envs}{RasterStack of environmental data}

\item{bg}{A data.frame of background points. This should have two columns with the column names matching the occurrence dataset.}

\item{tune.args}{A named list containing feature class settings ("fc") and regularization values ("rm"). Use ?ENMevaluate for details}

\item{partitions}{Character. The name of data partitioning strategy to use for model evaluations. Can select from: "randomkfold", "block", "checkerboard1", "checkerboard2", "testing", and "user"}

\item{partition.settings}{A named list used to specify settings for the selected partitioning scheme}

\item{user.grp}{A named list designating user-specified cross-validation folds. Only required when "partition.settings" is set to "user"}

\item{type}{Character. Select a method for selecting the optimal parameter combinations. If "type1" is used, the optimal model is selected based on the minimum 10% omission rate as a primary criterion, minimum AUCDIFF value as a secondary criterion, and maximum validation AUC as a tertiary criterion
If "type2" is used, the primary, secondary, and tertiary criteria for model selection would be delta.AICc smaller than 2, minimum 10% omission rate, and maximum validation AUC, respectively}
}
\value{
A named list with slots for model metrics of selected models, model objects, model predictions, and variable importance
}
\description{
MaxEnt model tuning for multiple species
}
\examples{
test_sp <- test_multisp(taxon.list = c('sp1', 'sp2', 'sp3'), occs.list = list(occs1, occs2, occs3), envs = envs, bg = bg, tune.args = tune.args, partitions = 'randomkfold', partition.settings = list(kfolds = 10), type = 'type1')
}
