% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitCheck.R
\name{unitCheck}
\alias{unitCheck}
\title{Check the units of calibration and projection data, using the maximum value of extracted raster cell values}
\usage{
unitCheck(ref.env, proj.env, n)
}
\arguments{
\item{ref.env}{RasterStack object of calibration data}

\item{proj.env}{RasterStack object of projection data}

\item{n}{Numeric. The number of points to use to sample the input raster cells. This argument uses the "randomPoints" function of the dismo package}
}
\value{
A data.frame object containing the maximum extracted cell values for calibration and projection data, respectively
}
\description{
Check the units of calibration and projection data, using the maximum value of extracted raster cell values
}
\examples{
a <- unitCheck(ref.env = ref.stack, proj.env = proj.stack, n = 10000)
}
