% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predictr.R
\name{model_predictr}
\alias{model_predictr}
\title{Predict ecological niche models to multiple datasets}
\usage{
model_predictr(model, preds.list, pred.names, method)
}
\arguments{
\item{model}{A fitted ENM of the class MaxEnt. Can be extracted by accessing the "models" slot output from "test_multibg" or "test_multisp". If "method = multi2single", this object will be of the class "list"}

\item{preds.list}{A list containing RasterStack objects of projection datasets. If "method = multi2single", input RasterStack object}

\item{pred.names}{A vector of names corresponding to each dataset in "preds.list". Can be the name of a particular time period (e.g. LGM) or climate change scenario (e.g. SSP26)}

\item{method}{Specify the type of model prediction. If "single2multi", conduct model predictions to multiple past/future environment from a single model. If "multi2single", make predictions to a single past/future dataset from multiple current models}
}
\value{
RasterStack of ENM predictions
}
\description{
Predict ecological niche models to multiple datasets
}
\examples{
pred.singe2multi <- model_predictr(model = mx_model, preds.list = list(lgm, mh, current), preds.names = c('LGM', 'Mid-Holocene', 'Current'), method = 'single2multi')
pred.multi2single <- model_predictr(model = mx_models_multi, preds.list = future.envs, preds.names = 'SSP585', method = 'multi2single')
}
