% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_response.R
\name{plot_response}
\alias{plot_response}
\title{Plot MaxEnt response curves}
\usage{
plot_response(
  resp.data,
  ncol = NULL,
  nrow = NULL,
  color = "cornflowerblue",
  linewidth = 1.2,
  alpha = 0.4
)
}
\arguments{
\item{resp.data}{A data.frame object output from "respDataPull" function}

\item{ncol}{The number of columns to use for ggplot2 facetting. Default is NULL}

\item{nrow}{The number of rows to use for ggplot2 facetting. Default is NULL}

\item{color}{Color palette to use for response curves}

\item{linewidth}{Line width for response curves. Default is 1.2}

\item{alpha}{Transparency. Default is 0.4}
}
\value{
A ggplot2-style facetted plot of the MaxEnt response curves
}
\description{
Plot MaxEnt response curves
}
\examples{
resp.plot <- plot_response(resp.data = resp.data, ncol = 2, nrow = 3, color = 'cornflowerblue', linewidth = 1.2, alpha = 0.4)
}
